+++ Changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file
 
         |     Option    	|  Default   	|                  Description                                    	|
         | :-------------	| :--------: 	| :-----------------------------------------                      	|
+        | Enabled       	| `True`     	|  Enable/Disable this plugin                                     	|
         | SizeDivisor   	| `1024`     	|  File size format: decimal (1k=1000=10³) or binary (1k=1024=2¹⁰)	|
         | MaxGlob       	| `5000`     	|  Skip folders with as many items                                	|
         | SymbolPane    	| `◧` `◨`    	|  `Left`/`Right` pane symbol                                     	|