+++ Changelog.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file
         | :-------------	| :--------: 	| :-----------------------------------------                      	|
         | Enabled       	| `True`     	|  Enable/Disable this plugin                                     	|
         | SizeDivisor   	| `1024`     	|  File size format: decimal (1k=1000=10³) or binary (1k=1024=2¹⁰)	|
-        | MaxGlob       	| `5000`     	|  Skip folders with as many items                                	|
+        | MaxGlob       	| `5000`     	|  Skip folders with as many items (folders+files)                	|
         | SymbolPane    	| `◧` `◨`    	|  `Left`/`Right` pane symbol                                     	|
         | SymbolHiddenF 	| `◻` `◼`    	|  Hidden files `Shown`/`Hidden` symbol                           	|
         | HideDotfile   	| `False`    	|  Treat .dotfiles as hidden files on Windows                     	|