+++ Changelog.md
@@ -2,7 +2,10 @@
 All notable changes to this project will be documented in this file
 
 ## [Unreleased]
-  [ Unreleased]: https://github.com/kek91/StatusBarExtended/compare/v0.3.1...HEAD
+  [ Unreleased]: https://github.com/kek91/StatusBarExtended/compare/v0.4.0...HEAD
+
+## [v0.4.0 — 13.08.2021]
+  [ v0.4.0 — 13.08.2021]: https://github.com/kek91/StatusBarExtended/releases/tag/v0.4.0
   - __Added__
     + :sparkles: User-configurable options via the `configure_status_bar_extended` command aliased as `StatusBarExtended: configure` in the Command Palette and bound to <kbd>Shift</kbd><kbd>F3</kbd> by default